Update so resolver is in IContainerRuntimeBase instead
