+++ packages/runtime/container-runtime-definitions/src/index.ts
@@ -45,4 +45,4 @@ export type {
 	IVersionMarkResolver,
 	ResolveResult,
 	VersionMarkCapture,
-} from "./versionMarks.js";
+} from "@fluidframework/runtime-definitions/internal";
